requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Pillow==10.1.0
gunicorn==20.1.0
jieba==0.42.1
python-dotenv==1.0.0

# LLM相关依赖
transformers>=4.39.0
torch>=2.2.0
accelerate>=0.27.0
bitsandbytes>=0.43.0
sentencepiece>=0.1.99
protobuf>=4.25.1
requests>=2.31.0 